requirements.txt
whisper
pydub
xattr
PyYAML
openai
